+++ tools/generate_shl.sh
@@ -1,0 +1,1 @@
+python shl.py ../src/include.hpp --header_search_paths ../src/ -o ../include/chess.hpp
